parens [nfc]: Simplify firstInStatement a bit
As it was, if there was a non-`n.Node` value in the middle of the
stack (is that even possible?), then we'd recheck the same parent
and child as we did on the last iteration.  That can't do anything
useful; instead, just have each iteration always check either the
current parent and child, or nothing.